% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_count.R
\name{grouped_count}
\alias{grouped_count}
\title{Get absolute frequency of a variable grouped by another variable}
\usage{
grouped_count(data, variable, grouping_variable)
}
\arguments{
\item{data}{Dataset containing the variables.}

\item{variable}{Character vector with variable name to be counted.}

\item{grouping_variable}{Character vector with grouping variable name.}
}
\value{
A tibble with the absolute frequency of the chosen \code{variable} grouped by the \verb{grouping variable}
}
\description{
Get absolute frequency of a variable grouped by another variable
}
