% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_three_waves.R
\name{plot_three_waves}
\alias{plot_three_waves}
\title{Plot numerical or categorical variables based on three-wave data}
\usage{
plot_three_waves(data, variable)
}
\arguments{
\item{data}{Dataset containing the variable to be plotted.}

\item{variable}{Variable to be plotted in a line or column plot. Numerical variables are plotted based on the mean.}
}
\value{
A ggplot2 plot using the input variable.
}
\description{
Plot numerical or categorical variables based on three-wave data
}
