% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_to_long.R
\name{transform_to_long}
\alias{transform_to_long}
\title{Transform variable to long format}
\usage{
transform_to_long(data, variable)
}
\arguments{
\item{data}{Dataset that contains variable to be transformed.}

\item{variable}{Variable to be transformed to long format.}
}
\value{
Tibble with input variable transformed to a long format.
}
\description{
Transform variable to long format
}
